+++ src/main/resources/inlayProviders/python.inlay.parameters/hints.classes.parameters.py
@@ -1,0 +1,7 @@
+class Person:
+    def __init__(self, name: str, age: int):
+        self.name = name
+        self.age = age
+
+
+Person("Alice", 17)