+++ src/main/resources/inlayProviders/python.inlay.parameters/hints.lambda.parameters.py
@@ -1,0 +1,3 @@
+lambda_exp = lambda char, times: char * times
+
+lambda_exp('a', 3)